feat: add match setup scene
